feat(gh-23): Move files creating logic to separate Service
