+++ src/main.rs
@@ -1,13 +1,15 @@
-use anyhow::{anyhow, Context, Result};
+use anyhow::{Context, Result};
 use clap::{ArgEnum, Parser};
 use log::info;
 use reqwest::blocking::Client as ReqwestClient;
-use std::fs;
-use std::path::Path;
 
 mod contracts;
+
 use contracts::client::{Client, Request};
-use contracts::service::{Contracts, Service};
+use contracts::service::{Contracts, Service as ContractsService};
+
+mod files;
+use files::{Files, Service as FilesService};
 
 pub const CONTRACTS_DEST_DIR: &str = "./contracts";
 
@@ -94,54 +96,18 @@ fn main() -> Result<()> {
 
     let contracts_client = Client::new(args.api_key.as_str(), api_url, http_client);
 
-    let contracts_service = Service::new(&contracts_client);
+    let contracts_service = ContractsService::new(&contracts_client);
+
+    let files_service = FilesService::new();
 
     let contracts = contracts_service
         .get_contracts(&args.contract_address)
         .context("contracts_service failed to get contracts")?;
 
-    for c in contracts.iter() {
-        let mut p = Path::new(&c.path);
-
-        // some contracts path contain a leading "/",
-        // make sure it's removed before creating the file
-        if p.starts_with("/") {
-            p = p.strip_prefix("/").unwrap();
-        }
-
-        // make sure the file has '.sol' extension
-        if p.file_name().is_none() {
-            return Err(anyhow!("failed to get file name from path: {:?}", p));
-        }
-
-        let mut file_name = p.file_name().unwrap().to_str().unwrap();
-        let mut tmp: String;
-        if p.extension().is_none() {
-            tmp = String::from(file_name);
-            tmp.push_str(".sol");
-            file_name = tmp.as_str();
-        }
-
-        // create all the dirs in Solidity file's path
-        if p.parent().is_none() {
-            return Err(anyhow!(
-                "failed to get path without file name from path: {:?}",
-                p
-            ));
-        }
-
-        let contract_dir = Path::new(&args.files_dest_path)
-            .join(network_name)
-            .join(&args.contract_address)
-            .join(p.parent().unwrap());
-
-        fs::create_dir_all(&contract_dir).context("failed to create contracts dir")?;
-
-        // create Solidity file
-        let contract_file_path = contract_dir.join(file_name);
-        fs::write(contract_file_path, &c.code)
-            .context("failed to crate and write to solidity file")?;
-    }
-
-    Ok(())
+    files_service.create_contract_files(
+        &args.files_dest_path,
+        network_name,
+        &args.contract_address,
+        contracts,
+    )
 }
